Separate Value from Critic
