+++ tests/policies/gac/test_Critic.py
@@ -1,9 +1,9 @@
-from policies.gac.gac_networks import Value, Critic
-from policies.policy_helpers.helpers import Transition
+from policies.gac.networks import Value, Critic
+from policies.policy_helpers.helper_classes import Transition
 import tensorflow as tf
 
 critic = Critic(3+4, 2)
-value = Value(3, 1)
+value = Value(3)
 
 transitions = Transition(
         tf.convert_to_tensor(
